Fix calculating dependencies hash on different machines or project clones
I noticed that `tuist cache print-hashes` was returning different values for the same branch and commit, in two different machines. I cloned the same repo in my own computer, and noticed that the hash was different - debugging took me to `DependencyContentHasher`, which used the absolute path, instead of the relative one.

Tests were updated accordingly.

This should fix a lot of stuff for remote caching, where many developers will calculate the hashes in their own computers. :)